fix: 共通化できそうなvalidate methodを切り出す
